pyproject.toml
[project]
name = "heartbeat-monitor"
version = "0.1.0"
description = "Heartbeat monitoring system with email alerts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "sqlalchemy>=2.0.0",
    "apscheduler>=3.10.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "email-validator>=2.0.0",
    "pytest>=7.4.0",
    "httpx>=0.24.0",
]

[tool.setuptools]
packages = ["server", "client"]

[project.scripts]
hb-agent = "client.cli:main"
